% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSpredict.R
\name{PLSpredict}
\alias{PLSpredict}
\title{PLSpredict performs predictions on SEMinR PLS models}
\usage{
PLSpredict(model = seminr_model, testData = test_set, technique = predict_DA)
}
\arguments{
\item{model}{A SEMinR model that has been trained on the training data}

\item{testData}{The testing data for which predictions will be generated}

\item{technique}{The predictive technique to be employed, Earliest Antecedents (EA) \code{predict_EA} or
Direct Antecedents (DA) \code{predict_DA}}
}
\description{
\code{PLSpredict} performs predictions on PLS models generated by SEMinR.
}
\details{
This function performs construct and item level predictions
}
\examples{
data(mobi)
trainData <- mobi[1:200,]
testData <- mobi[201:250,]
library(seminr)

# seminr syntax for creating measurement model
mobi_mm <- constructs(
  composite("Image",        multi_items("IMAG", 1:5)),
  composite("Expectation",  multi_items("CUEX", 1:3)),
  composite("Value",        multi_items("PERV", 1:2)),
  composite("Satisfaction", multi_items("CUSA", 1:3))
)

mobi_sm <- relationships(
  paths(to = "Satisfaction",
        from = c("Image", "Expectation", "Value"))
)

mobi_train <- estimate_pls(trainData, mobi_mm, interactions = NULL, mobi_sm)
predictions <- PLSpredict(mobi_pls, testData)

}
