% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePredict.R
\name{validatePredict}
\alias{validatePredict}
\title{PLSpredict performs predictions on SEMinR PLS models}
\usage{
validatePredict(model = seminr_model, technique = predict_DA, noFolds = 10)
}
\arguments{
\item{model}{A SEMinR model that has been trained on the training data}

\item{technique}{The predictive technique to be employed, Earliest Antecedents (EA) \code{predict_EA} or
Direct Antecedents (DA) \code{predict_DA}}

\item{noFolds}{The number of folds to perform. k in k-fold. Default is 10.}
}
\description{
\code{validatePredict} performs k-fold cross validation and generates predictive metrics for PLS models generated by SEMinR.
}
\details{
This function performs k-fold cross validation and generates RMSE, MAPE and MAD for the PLS model and
LM benchmark.
}
\examples{
data(mobi)
library(seminr)

# seminr syntax for creating measurement model
mobi_mm <- constructs(
  composite("Image",        multi_items("IMAG", 1:5)),
  composite("Expectation",  multi_items("CUEX", 1:3)),
  composite("Value",        multi_items("PERV", 1:2)),
  composite("Satisfaction", multi_items("CUSA", 1:3))
)

mobi_sm <- relationships(
  paths(to = "Satisfaction",
        from = c("Image", "Expectation", "Value"))
)

mobi_pls <- estimate_pls(mobi, mobi_mm, interactions = NULL, mobi_sm)
metrics <- validatePredict(mobi_pls)

}
